+++ src/main/java/com/tortilla/f1/telemetry/telemetryproducer/model/Headers.java
@@ -6,23 +6,23 @@ import lombok.Data;
 @Data
 public class Headers {
     @JsonAlias("m_packetFormat")
-    private int packetFormat;
+    private Float packetFormat;
     @JsonAlias("m_gameMajorVersion")
-    private int gameMajorVersion;
+    private Float gameMajorVersion;
     @JsonAlias("m_gameMinorVersion")
-    private int gameMinorVersion;
+    private Float gameMinorVersion;
     @JsonAlias("m_packetVersion")
-    private int packetVersion;
+    private Float packetVersion;
     @JsonAlias("m_packetId")
-    private int packetId;
+    private Float packetId;
     @JsonAlias("m_sessionUID")
     private String sessionUID;
     @JsonAlias("m_sessionTime")
     float sessionTime;
     @JsonAlias("m_frameIdentifier")
-    private int frameIdentifier;
+    private Float frameIdentifier;
     @JsonAlias("m_playerCarIndex")
-    private int playerCarIndex;
+    private Float playerCarIndex;
     @JsonAlias("m_secondaryPlayerCarIndex")
-    private int secondaryPlayerCarIndex;
+    private Float secondaryPlayerCarIndex;
 }
